+++ src/main/java/wooteco/subway/service/member/FavoriteRequest.java
@@ -1,7 +1,7 @@
-package wooteco.subway.service.member.favorite.dto;
+package wooteco.subway.service.member;
 
 
-import wooteco.subway.domain.member.favorite.Favorite;
+import wooteco.subway.domain.member.Favorite;
 
 public class FavoriteRequest {
     private String source;
