+++ src/test/java/wooteco/subway/acceptance/member/FavoriteAcceptanceTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.acceptance.AcceptanceTest;
+import wooteco.subway.service.member.FavoriteResponse;
 import wooteco.subway.service.member.dto.TokenResponse;
-import wooteco.subway.service.member.favorite.dto.FavoriteResponse;
 
 import java.util.HashMap;
 import java.util.List;
