+++ src/test/java/wooteco/subway/domain/member/MemberTest.java
@@ -2,7 +2,6 @@ package wooteco.subway.domain.member;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import wooteco.subway.domain.member.favorite.Favorite;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
