+++ src/test/java/wooteco/subway/web/member/FavoriteControllerTest.java
@@ -17,9 +17,9 @@ import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.filter.CharacterEncodingFilter;
 import org.springframework.web.filter.ShallowEtagHeaderFilter;
 import wooteco.subway.docs.FavoriteDocumentation;
-import wooteco.subway.service.member.favorite.FavoriteService;
-import wooteco.subway.service.member.favorite.dto.FavoriteRequest;
-import wooteco.subway.service.member.favorite.dto.FavoriteResponse;
+import wooteco.subway.service.member.FavoriteRequest;
+import wooteco.subway.service.member.FavoriteResponse;
+import wooteco.subway.service.member.FavoriteService;
 
 import java.util.Arrays;
